+++ QianBao/Classes/AddPayViewController.swift
@@ -283,7 +283,7 @@ class AddPayViewController: UIViewController,ZYKeyboardDelegate,UITextFieldDeleg
         }
         let money:String =  self.money.text!.replacingOccurrences(of: "￥", with: "")
         let desc:String = self.desctext.text!
-        if Int(money) == nil {
+        if Float(money) == nil {
             alertController = UIAlertController(title: "",message: "金额不正确", preferredStyle: .alert)
             alertController.addAction(okAction)
             self.present(alertController, animated: true, completion: nil)
